+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Release 5.5.0:
    - Update `signum` to 3.15.0, supporting X.509 certificates in v1, v2 too
    - Update JsonPath4K
  - Introduce dedicated Android targets, separate from JVM targets, that compileto JDK 8 / API-Level 30
+ - Improved error logging and exposing for presentation exchange input evaluation
 
 Release 5.4.3:
  - Fix property names for serialized RQES data classes